Fix tests to better reflect asynchronous behavior of webpack/requirejs internals.
The AMD module in `amd-exec.js` requires `require-packages.js` which in turn
require.ensures `packages.js`.  However, the `amd-exec.js` module will not
wait for `packages.js` to load since it is not an explicit dependency.  For
that reason, the load order cannot be guaranteed - instead, simply test that
both the webpack module and the AMD module load successfully.